future(upload): small chores (#25470)
