+++ packages/core/upload/admin/src/future/pages/Assets/AssetsPage.tsx
@@ -1,7 +1,7 @@
-import { useEffect, useRef, useCallback, useState, type ChangeEvent } from 'react';
+import { useRef, useCallback, type ChangeEvent } from 'react';
 
 import * as ToggleGroup from '@radix-ui/react-toggle-group';
-import { Layouts, SearchInput, useElementOnScreen } from '@strapi/admin/strapi-admin';
+import { Layouts, useElementOnScreen, usePersistentState } from '@strapi/admin/strapi-admin';
 import {
   Box,
   Flex,
@@ -11,21 +11,17 @@ import {
   Typography,
   VisuallyHidden,
 } from '@strapi/design-system';
-import { ChevronDown, Files, Folder, GridFour as GridIcon, List } from '@strapi/icons';
+import { ChevronDown, Files, GridFour as GridIcon, List } from '@strapi/icons';
 import { useIntl } from 'react-intl';
 import { styled } from 'styled-components';
 
-import { usePersistentState } from '../../../hooks/usePersistentState';
 import { useUploadFilesStreamMutation } from '../../services/api';
 import { getTranslationKey } from '../../utils/translations';
 
 import { AssetsGrid } from './components/AssetsGrid';
 import { AssetsTable } from './components/AssetsTable';
-import { DropZoneWithOverlay } from './components/DropZone/DropZoneWithOverlay';
-import {
-  UploadDropZoneProvider,
-  useUploadDropZone,
-} from './components/DropZone/UploadDropZoneContext';
+import { DropFilesMessage, DropZoneWithOverlay } from './components/DropZone/UploadDropZone';
+import { UploadDropZoneProvider } from './components/DropZone/UploadDropZoneContext';
 import { localStorageKeys, viewOptions } from './constants';
 import { useInfiniteAssets } from './hooks/useInfiniteAssets';
 
@@ -136,72 +132,14 @@ const StyledToggleItem = styled(ToggleGroup.Item)`
   }
 `;
 
-/**
- * Dropzone items
- */
+const HeaderWrapper = styled.div`
+  [data-strapi-header] {
+    background: ${({ theme }) => theme.colors.neutral0};
 
-interface DropFilesMessageProps {
-  uploadDropZoneRef?: React.RefObject<HTMLDivElement>;
-}
-
-const DropFilesMessage = ({ uploadDropZoneRef }: DropFilesMessageProps) => {
-  const { formatMessage } = useIntl();
-  const { isDragging } = useUploadDropZone();
-
-  // Dropzone message position (relative to main content)
-  const [leftContentWidth, setLeftContentWidth] = useState(0);
-
-  // Calculate the left content width to position the dropzone message correctly
-  useEffect(() => {
-    if (!uploadDropZoneRef?.current) return;
-
-    const updateRect = () => {
-      const rect = uploadDropZoneRef.current?.getBoundingClientRect();
-      if (rect) {
-        setLeftContentWidth((prev) => (prev !== rect.left ? rect.left : prev));
-      }
-    };
-
-    updateRect();
-    const resizeObserver = new ResizeObserver(updateRect);
-    resizeObserver.observe(uploadDropZoneRef.current);
-    return () => resizeObserver.disconnect();
-  }, [uploadDropZoneRef]);
-
-  if (!isDragging) return null;
-
-  return (
-    <DropFilesMessageImpl $leftContentWidth={leftContentWidth}>
-      <Typography textColor="neutral0">
-        {formatMessage({
-          id: getTranslationKey('dropzone.upload.message'),
-          defaultMessage: 'Drop here to upload to',
-        })}
-      </Typography>
-      <Flex gap={2} alignItems="center">
-        <Folder width={20} height={20} fill="neutral0" />
-        <Typography textColor="neutral0" fontWeight="semiBold">
-          Current folder{/* TODO: Replace this later with the current folder name */}
-        </Typography>
-      </Flex>
-    </DropFilesMessageImpl>
-  );
-};
-
-const DropFilesMessageImpl = styled(Box)<{ $leftContentWidth: number }>`
-  position: fixed;
-  bottom: ${({ theme }) => theme.spaces[8]};
-  left: 50%;
-  transform: translateX(calc(-50% + ${({ $leftContentWidth }) => $leftContentWidth / 2}px));
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: ${({ theme }) => theme.spaces[2]};
-  background: ${({ theme }) => theme.colors.primary600};
-  padding: ${({ theme }) => theme.spaces[4]} ${({ theme }) => theme.spaces[6]};
-  border-radius: ${({ theme }) => theme.borderRadius};
-  z-index: 2;
+    h1 {
+      font-size: 1.8rem;
+    }
+  }
 `;
 
 export const AssetsPage = () => {
@@ -261,17 +199,15 @@ export const AssetsPage = () => {
   return (
     <UploadDropZoneProvider onDrop={handleDrop}>
       <Box ref={uploadDropZoneRef}>
-        <Layouts.Root minHeight="100vh">
+        <Layouts.Root minHeight="100vh" background="neutral0">
           <VisuallyHidden>
             <input type="file" ref={fileInputRef} onChange={handleFileChange} multiple />
           </VisuallyHidden>
 
-          <DropFilesMessage uploadDropZoneRef={uploadDropZoneRef} />
-
-          <Layouts.Header
-            title="TODO: Folder location"
-            primaryAction={
-              <Flex gap={2}>
+          <HeaderWrapper>
+            <Layouts.Header
+              title="TODO: Folder name"
+              primaryAction={
                 <SimpleMenu
                   popoverPlacement="bottom-end"
                   variant="default"
@@ -285,58 +221,62 @@ export const AssetsPage = () => {
                     })}
                   </MenuItem>
                 </SimpleMenu>
-                <SearchInput
-                  label={formatMessage({
-                    id: getTranslationKey('search.label'),
-                    defaultMessage: 'Search for an asset',
-                  })}
-                  trackedEvent="didSearchMediaLibraryElements"
-                  trackedEventDetails={{ location: 'upload' }}
-                />
-                <StyledToggleGroup
-                  type="single"
-                  value={isGridView ? 'grid' : 'table'}
-                  onValueChange={(value) =>
-                    value && setView(value === 'grid' ? viewOptions.GRID : viewOptions.TABLE)
-                  }
-                  aria-label={formatMessage({
-                    id: getTranslationKey('view.switch.label'),
-                    defaultMessage: 'View options',
-                  })}
-                >
-                  <StyledToggleItem
-                    value="table"
-                    aria-label={formatMessage({
-                      id: getTranslationKey('view.table'),
-                      defaultMessage: 'Table view',
-                    })}
-                  >
-                    <List />
-                    {formatMessage({
-                      id: getTranslationKey('view.table'),
-                      defaultMessage: 'Table view',
-                    })}
-                  </StyledToggleItem>
-                  <StyledToggleItem
-                    value="grid"
-                    aria-label={formatMessage({
-                      id: getTranslationKey('view.grid'),
-                      defaultMessage: 'Grid view',
-                    })}
-                  >
-                    <GridIcon />
-                    {formatMessage({
-                      id: getTranslationKey('view.grid'),
-                      defaultMessage: 'Grid view',
-                    })}
-                  </StyledToggleItem>
-                </StyledToggleGroup>
-              </Flex>
-            }
-          />
+              }
+              subtitle={
+                <Flex justifyContent="space-between" alignItems="center" gap={4} width="100%">
+                  <Flex gap={4} alignItems="center">
+                    TODO: Filters and search
+                  </Flex>
+
+                  <Flex gap={4} alignItems="center">
+                    <Box>TODO: Sort</Box>
+                    <StyledToggleGroup
+                      type="single"
+                      value={isGridView ? 'grid' : 'table'}
+                      onValueChange={(value) =>
+                        value && setView(value === 'grid' ? viewOptions.GRID : viewOptions.TABLE)
+                      }
+                      aria-label={formatMessage({
+                        id: getTranslationKey('view.switch.label'),
+                        defaultMessage: 'View options',
+                      })}
+                    >
+                      <StyledToggleItem
+                        value="table"
+                        aria-label={formatMessage({
+                          id: getTranslationKey('view.table'),
+                          defaultMessage: 'Table view',
+                        })}
+                      >
+                        <List />
+                        {formatMessage({
+                          id: getTranslationKey('view.table'),
+                          defaultMessage: 'Table view',
+                        })}
+                      </StyledToggleItem>
+                      <StyledToggleItem
+                        value="grid"
+                        aria-label={formatMessage({
+                          id: getTranslationKey('view.grid'),
+                          defaultMessage: 'Grid view',
+                        })}
+                      >
+                        <GridIcon />
+                        {formatMessage({
+                          id: getTranslationKey('view.grid'),
+                          defaultMessage: 'Grid view',
+                        })}
+                      </StyledToggleItem>
+                    </StyledToggleGroup>
+                  </Flex>
+                </Flex>
+              }
+            />
+          </HeaderWrapper>
 
           <Layouts.Content>
             <DropZoneWithOverlay>
+              <DropFilesMessage uploadDropZoneRef={uploadDropZoneRef} />
               <AssetsView view={view} />
             </DropZoneWithOverlay>
           </Layouts.Content>
